+++ core/src/integration-test/java/com/scalar/db/api/TwoPhaseCommitTransactionIntegrationTestBase.java
@@ -10,7 +10,6 @@ import com.google.common.collect.ImmutableSet;
 import com.scalar.db.api.GetBuilder.BuildableGet;
 import com.scalar.db.api.Scan.Ordering;
 import com.scalar.db.api.ScanBuilder.BuildableScanOrScanAllFromExisting;
-import com.scalar.db.config.DatabaseConfig;
 import com.scalar.db.exception.storage.ExecutionException;
 import com.scalar.db.exception.transaction.CommitException;
 import com.scalar.db.exception.transaction.CrudException;
@@ -24,6 +23,7 @@ import com.scalar.db.io.Key;
 import com.scalar.db.io.Value;
 import com.scalar.db.service.TransactionFactory;
 import com.scalar.db.transaction.consensuscommit.Attribute;
+import com.scalar.db.transaction.consensuscommit.ConsensusCommitConfig;
 import com.scalar.db.transaction.consensuscommit.ConsensusCommitUtils;
 import com.scalar.db.util.TestUtils;
 import com.scalar.db.util.TestUtils.ExpectedResult;
@@ -83,7 +83,7 @@ public abstract class TwoPhaseCommitTransactionIntegrationTestBase {
     manager = factory.getTwoPhaseCommitTransactionManager();
 
     Properties debugProperties = TestUtils.addSuffix(getProperties(), testName);
-    debugProperties.setProperty(DatabaseConfig.DEBUG, "true");
+    debugProperties.setProperty(ConsensusCommitConfig.DEBUG, "true");
     managerWithDebug =
         TransactionFactory.create(debugProperties).getTwoPhaseCommitTransactionManager();
   }
@@ -1071,7 +1071,7 @@ public abstract class TwoPhaseCommitTransactionIntegrationTestBase {
     selection_UsingDebugMode_ShouldReturnCorrectColumns(false, true);
   }
 
-  public void selection_UsingDebugMode_ShouldReturnCorrectColumns(
+  private void selection_UsingDebugMode_ShouldReturnCorrectColumns(
       boolean isScan, boolean hasProjections) throws TransactionException {
     // Arrange
     Put put =
